third_party/WebKit/LayoutTests/fast/dom/Document/createCDATASection.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
</head>
<body>
<script>
window.xmlDoc = document.implementation.createDocument(null, "example", null);
shouldBe('"createCDATASection" in window.xmlDoc', 'true');
shouldBeEqualToString('window.xmlDoc.createCDATASection("Hello").data', "Hello");
shouldThrow('window.xmlDoc.createCDATASection("Ha]]>o")');
shouldBeEqualToString('window.xmlDoc.createCDATASection(null).data', "null");
</script>
</body>
</html>
